\name{riboSeq-package}
\alias{riboSeq-package}
\alias{riboSeq}
\docType{package}
\title{
Analysis of sequencing data from ribosome profiling experiments.
}
\description{
Plotting functions, frameshift detection and parsing of sequencing
    data from ribosome profiling experiments.
}
\details{
\tabular{ll}{
Package: \tab riboSeq\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2014-05-02\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 3.0.2), methods, baySeq, GenomicRanges, abind\cr
Imports: \tab GenomicRanges, methods, abind\cr
Packaged: \tab 2014-05-16 10:49:24 UTC; tjh48\cr
Built: \tab R 3.0.2; ; 2014-05-16 10:49:44 UTC; unix\cr
}

Further information is available in the following vignettes:
\tabular{ll}{
\code{riboSeq} \tab riboSeq (source, pdf)\cr
}
}

\author{
Thomas J. Hardcastle

Maintainer: Thomas J. Hardcastle <tjh48@cam.ac.uk>
}

\keyword{ package }
